Re-add input_helpers.py, split common.py
