+++ src/dq_suite/__init__.py
@@ -1,7 +1,7 @@
 """DQ API."""
 
-from src.dq_suite.common import export_schema
 from src.dq_suite.df_checker import ValidationSettings, run
+from src.dq_suite.input_helpers import schema_to_json_string
 
 # Use __all__ to let developers know what is part of the public API.
-__all__ = ["export_schema", "ValidationSettings", "run"]
+__all__ = ["schema_to_json_string", "ValidationSettings", "run"]
